requirements.txt
fastapi
uvicorn
httpx
pyyaml
slowapi